Added room assignment event handler
