+++ backend/src/index.ts
@@ -26,7 +26,6 @@ server.listen(PORT, () => {
 // the new socket should send the username of the client through the auth object
 io.on("connection", (socket: Socket) => {
   const username = socket.handshake.auth.username;
-  let currRoom = "Room " + boards.size;
 
   // Check if user is already in a room
   const userRoomData = rooms.get(username);
@@ -34,26 +33,9 @@ io.on("connection", (socket: Socket) => {
     // If so, place socket back into that room
     const newSocket: RoomData = { room: userRoomData.room, socket: socket.id };
     rooms.set(username, newSocket);
-  } else {
-    let board: BoardState = { lock: 0, board: "", players: [username] };
-
-    // socket only has default socket ID room => assign socket to a room
-    if (socket.rooms.size == 1) {
-      if (io.sockets.adapter.rooms.get(currRoom)) {
-        //Check if current room is full => create new room
-        if (io.sockets.adapter.rooms.get(currRoom).size == 2) {
-          currRoom = "Room " + (boards.size + 1);
-        } else {
-          const currBoard = boards.get(currRoom);
-          if (currBoard) board = { ...board, players: [currBoard.players[0], username] };
-        }
-      }
-      boards.set(currRoom, board);
-      rooms.set(username, { room: currRoom, socket: socket.id });
-      socket.join(currRoom);
-    }
   }
 
+  require("./eventHandlers/roomHandler.ts")(socket, io, username, rooms, boards);
   require("./eventHandlers/moveHandler.ts")(socket, io, username, rooms, boards);
   require("./eventHandlers/resignHandler.ts")(socket, io, username, rooms, boards);
   require("./eventHandlers/drawHandler.ts")(socket, io, username, rooms, boards);
